+++ Classes/Domain/Converter/DatasetConverter.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 
 namespace Brotkrueml\JobRouterData\Domain\Converter;
 
+use Brotkrueml\JobRouterData\Domain\Model\Column;
 use Brotkrueml\JobRouterData\Domain\Model\Table;
 use Brotkrueml\JobRouterData\Event\ModifyColumnContentEvent;
 use Psr\EventDispatcher\EventDispatcherInterface;
@@ -31,7 +32,7 @@ final class DatasetConverter
     }
 
     /**
-     * @return list<array<string, float|int|string|null>>
+     * @return list<array<string, float|int|string>>
      */
     public function convertFromJsonToArray(Table $table, string $locale): array
     {
@@ -41,15 +42,15 @@ final class DatasetConverter
 
             $row = [];
             foreach ($table->getColumns() as $column) {
-                if ($datasetArray[$column->getName()] ?? false) {
-                    /** @var float|int|string|null $content */
-                    $content = $datasetArray[$column->getName()];
-                    if ($content === null || $content === '') {
-                        $row[$column->getName()] = '';
+                $columnName = $column->getName();
+                if ($datasetArray[$columnName] ?? false) {
+                    $row['_original_' . $columnName] = $datasetArray[$column->getName()];
+                    if ($row['_original_' . $columnName] === null || $row['_original_' . $columnName] === '') {
+                        $row[$columnName] = '';
                         continue;
                     }
 
-                    $event = new ModifyColumnContentEvent($table, $column, $content, $locale);
+                    $event = new ModifyColumnContentEvent($table, $column, $row['_original_' . $columnName], $locale);
                     /** @var ModifyColumnContentEvent $event */
                     $event = $this->eventDispatcher->dispatch($event);
                     $row[$column->getName()] = $event->getContent();
@@ -58,6 +59,49 @@ final class DatasetConverter
             $rows[] = $row;
         }
 
+        return $this->sortRowsByColumns($table->getColumns()->toArray(), $rows);
+    }
+
+    /**
+     * @param Column[] $columns
+     * @param list<array<string, float|int|string>> $rows
+     * @return list<array<string, float|int|string>>
+     */
+    public function sortRowsByColumns(array $columns, array $rows): array
+    {
+        $columnsToSortBy = $this->getSortingColumns($columns);
+        if ($columnsToSortBy !== []) {
+            \usort($rows, static function (array $a, array $b) use ($columnsToSortBy): int {
+                foreach ($columnsToSortBy as $column) {
+                    $order = $column->getSortingOrder() === 'desc' ? -1 : 1;
+                    $result = ($a['_original_' . $column->getName()] <=> $b['_original_' . $column->getName()]) * $order;
+                    if ($result !== 0) {
+                        return $result;
+                    }
+                }
+
+                return 0;
+            });
+        }
+
         return $rows;
+    }
+
+    /**
+     * @param Column[] $columns
+     * @return Column[]
+     */
+    public function getSortingColumns(array $columns): array
+    {
+        $columnsToSortBy = \array_values(
+            \array_filter($columns, static function (Column $column): bool {
+                return $column->getSortingPriority() > 0;
+            })
+        );
+        \usort($columnsToSortBy, static function (Column $a, Column $b): int {
+            return $a->getSortingPriority() <=> $b->getSortingPriority();
+        });
+
+        return $columnsToSortBy;
     }
 }