+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Extensible formatters for table columns in content element (#10)
-- Select alignment for table columns in content element (#11)
+- Formatters for table columns in content element (#10)
+- Alignment for table columns in content element (#11)
+- Sorting of table columns in content element (#12)
 - PSR-14 event to adjust or reject a dataset on synchronisation (#13)
 - Option in sync command for forcing a synchronisation